Refactor for models
